pyproject.toml
[tool.poetry]
name = "cowsway"
version = "0.1.0"
description = ""
authors = ["Imraj423 <dahqniss@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.3"
django-simple-history = "^2.8.0"
django-field-history = "^0.8.0"
django-material = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
